Extract shared command runner from check
